Added Polymorphic to TranslateData
